Remove deployment bucket stack from build scripts
Change scripts to bash

Make s3 removal stepwise

feat: Add helper scripts

refactor(package.json): upgrade commitlint

refactor(adding single deployment bucket feature): add Single deployment bucket

Update scripts

Update scripts

Add removal of files from deploymentbucket after deploy

Update README

feat: upgrade features project

Add missing update to yml

Add metric configuration

refactor(adding single deployment bucket feature): add Single deployment bucket

Fix Merge issue duplicate

fix: fix stackremoval using SLS remove instead of AWS CLI

fix: create deploymentBucket using serverless instead of AWS Cli

fix: rebase

fix: updated service list

fix: bug in s3remove script

+++ scripts/s3remove.sh
@@ -1,11 +1,37 @@
-#!/bin/zsh
-DIR=${0:a:h}
+#!/bin/bash
+DIR=$(dirname $0)
+
+if [ -z ${AWS_VAULT} ]; then
+     echo AWS Vault needs to be installed to run this script.
+     exit 1
+ fi
+
+ which aws >/dev/null 2>&1
+
+ if [ $? -ne 0 ]; then
+     echo AWS CLI needs to be installed to run this script.
+     exit 1
+ fi
+
+ clear
+ echo This script removes AWS S3 buckets from an AWS account.
+ echo Before running, make sure that you are logged on to the account
+ echo of which buckets you want to remove
+ echo
+ echo The currently active source account is \"${AWS_VAULT:Unknown}\"
+echo
 
 BUCKETS=(`aws s3 ls | cut -f3 -d' '`)
 
-for BUCKET in $BUCKETS
+for BUCKET in ${BUCKETS[@]}
 do
-    echo "Removing bucket: ${BUCKET}"
-    aws s3 rm s3://${BUCKET} --recursive
-    aws s3 rb s3://${BUCKET}
+    while read -r -n1 -p "Remove bucket ${BUCKET} [y/n]? " ANSWER;do
+        echo
+        if [ "${ANSWER}" == "y" ]; then
+            echo "Removing bucket: ${BUCKET}"
+            aws s3 rm s3://${BUCKET} --recursive
+            aws s3 rb s3://${BUCKET}
+        fi
+        break;
+    done
 done
